refactoring entire code base
