+++ .gitignore
@@ -2,10 +2,11 @@
 .venv/
 secrets/
 .idea/
-monkelib/__pycache__/
-monkelib/API/__pycache__/
-monkelib/Tests/__pycache__/
-monkelib/Tests/.pytest_cache/
-monkelib/Tests/functional/__pycache__/
-monkelib/Tests/functional/.pytest_cache/
-monkelib/PretrainedModels/
+codecompasslib/__pycache__/
+codecompasslib/API/__pycache__/
+codecompasslib/Tests/__pycache__/
+codecompasslib/Tests/.pytest_cache/
+codecompasslib/Tests/functional/__pycache__/
+codecompasslib/Tests/functional/.pytest_cache/
+codecompasslib/PretrainedModels/
+backend/__pycache__/
